requirements.txt
arcgis>=1.9.0,<1.10.0
pandas>=1.3.0,<1.4.0
python_dotenv>=0.19.0,<0.20.0
pyyaml>=5.4,<6.1
typer>=0.4.0,<0.5.0
docxtpl>=0.14.2,<0.16.0
pytz
pywin32
pillow
